Fixes for marketo routerTransform
